pyproject.toml
[project]
name = "promptme"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "openai==2.8.0",
    "pillow-heif==1.1.1",
    "python-dotenv==1.0.0",
    "streamlit==1.51.0",
    "streamlit-ace==0.1.1",
    "streamlit-chat==0.1.1",
]

[[tool.uv.index]]
url = "https://mirrors.aliyun.com/pypi/simple/"
default = true
